+++ server/src/main/java/vuatiengvietpj/controller/UserController.java
@@ -50,7 +50,7 @@ public class UserController extends ServerController {
     }
 
     public Response handleGetIn4(String data) {
-        return createSuccessResponse(module, "GETIN4", gson.toJson(userDAO.findById(Long.parseLong(data))));
+        return createSuccessResponse(module, "GETIN4", gson.toJson(userDAO.findById(Integer .parseInt(data))));
     }
 
     // check client còn sống k
@@ -89,7 +89,7 @@ public class UserController extends ServerController {
 
     // đăng xuất
     public Response handleLogOut(String data) {
-        Long userId = Long.parseLong(data);
+        Integer  userId = Integer .parseInt(data);
         boolean ok = SessionManager.destroy(userId); // hủy session
         return ok ? createSuccessResponse(module, "LOGOUT", "OK")
                 : createErrorResponse(module, "LOGOUT", "Not logged in");
